refactor(domain): using plain emotinon css and classNames
