+++ app/.eslintrc.js
@@ -43,7 +43,7 @@ module.exports = {
       'error',
       { devDependencies: ['**/*.stories.tsx'] },
     ],
-    '@typescript-eslint/no-use-before-define': ['error', { functions: false }],
+    '@typescript-eslint/no-use-before-define': ['off', { functions: false }],
     '@typescript-eslint/naming-convention': 'off',
     '@typescript-eslint/no-unused-vars': [
       'error',
