--- app/components/ContentContainer.tsx
@@ -1,21 +1,0 @@
-import styled from '@emotion/styled';
-import { Palette } from '../design/color';
-import { Breakpoints } from '../design/responsive';
-
-const { neutral, blue, contrast } = Palette;
-
-export const ContentContainer = styled.div`
-  background-color: ${neutral};
-  display: flex;
-  flex-direction: column;
-  margin: auto;
-  padding: 1rem;
-  border-radius: 1rem;
-  box-shadow: 0 0 2px 1px ${blue};
-  color: ${contrast};
-  max-width: ${Breakpoints.tabletPortrait}px;
-  text-align: center;
-  * {
-    color: ${contrast};
-  }
-`;
