+++ app/components/game/Card.tsx
@@ -5,16 +5,6 @@ import { Palette } from '../../design/color';
 import { beginAt, Breakpoints } from '../../design/responsive';
 import { CardColor, CardType, isSpyMaster, Player, Team } from '../../model';
 
-const { neutral, death, blue, red, contrast } = Palette;
-const { phoneLg, tabletPortrait } = Breakpoints;
-
-const CardColorMap: { [key in CardType]: string } = {
-  Neutral: neutral,
-  Death: death,
-  Blue: blue,
-  Red: red,
-};
-
 interface CardProps {
   card: CardColor;
   turn: Team;
@@ -34,28 +24,7 @@ export const Card: FC<CardProps> = ({
       type="button"
       onClick={onClick}
       disabled={isDisabled(turn, color, player)}
-      className={css`
-        background-color: ${color ? CardColorMap[color] : Palette.light};
-        box-shadow: 0 0 2px 1px ${Palette.blue};
-        color: ${color === 'Death' ? neutral : contrast};
-        border-radius: 0.25rem;
-        padding: 0.75rem 0.1rem;
-        word-break: break-all;
-        font-size: ${size}px;
-        cursor: ${isDisabled(turn, color, player) ? undefined : 'pointer'};
-        :hover {
-          background-color: ${isDisabled(turn, color, player)
-            ? undefined
-            : darken(0.1, Palette.light)};
-        }
-        ${beginAt(phoneLg)} {
-          font-size: ${size * 1.5}px;
-        }
-        ${beginAt(tabletPortrait)} {
-          font-size: ${size * 2}px;
-          padding: 1rem 0.1rem;
-        }
-      `}
+      className={styleButton(turn, color, size, player)}
     >
       {word}
     </button>
@@ -72,3 +41,46 @@ function isDisabled(
   if (isSpyMaster(player)) return true;
   return player.team !== turn;
 }
+
+function styleButton(
+  turn: Team,
+  color: CardType | null,
+  size: number,
+  player?: Player,
+): string {
+  return css`
+    background-color: ${color ? CardColorMap[color] : Palette.light};
+    box-shadow: 0 0 2px 1px ${Palette.blue};
+    color: ${color === 'Death' ? neutral : contrast};
+    border-radius: 0.25rem;
+    padding: 0.75rem 0.1rem;
+    word-break: break-all;
+    font-size: ${size}px;
+    cursor: ${isDisabled(turn, color, player) ? undefined : 'pointer'};
+
+    :hover {
+      background-color: ${isDisabled(turn, color, player)
+        ? undefined
+        : darken(0.1, Palette.light)};
+    }
+
+    ${beginAt(phoneLg)} {
+      font-size: ${size * 1.5}px;
+    }
+
+    ${beginAt(tabletPortrait)} {
+      font-size: ${size * 2}px;
+      padding: 1rem 0.1rem;
+    }
+  `;
+}
+
+const { neutral, death, blue, red, contrast } = Palette;
+const { phoneLg, tabletPortrait } = Breakpoints;
+
+const CardColorMap: { [key in CardType]: string } = {
+  Neutral: neutral,
+  Death: death,
+  Blue: blue,
+  Red: red,
+};
