+++ app/components/game/GuessLog.tsx
@@ -6,20 +6,7 @@ import { Palette } from '../../design/color';
 export type GuessLogProps = Pick<GameState, 'board'> & { guesses: number[] };
 
 export const GuessLog: FC<GuessLogProps> = ({ board, guesses }) => (
-  <div
-    className={css`
-      display: flex;
-      flex-direction: column;
-      color: ${Palette.neutral};
-      & h4 {
-        font-weight: bold;
-        margin: 0.5rem 0;
-      }
-      & li {
-        margin: 0.25rem 0;
-      }
-    `}
-  >
+  <div className={container}>
     <h4>Guess History</h4>
     <ul>
       {guesses
@@ -31,3 +18,16 @@ export const GuessLog: FC<GuessLogProps> = ({ board, guesses }) => (
     </ul>
   </div>
 );
+
+const container = css`
+  display: flex;
+  flex-direction: column;
+  color: ${Palette.neutral};
+  & h4 {
+    font-weight: bold;
+    margin: 0.5rem 0;
+  }
+  & li {
+    margin: 0.25rem 0;
+  }
+`;
