+++ app/components/game/PlayerList.tsx
@@ -1,64 +1,9 @@
 import React, { FC, useCallback } from 'react';
-import styled from '@emotion/styled';
 import { lighten } from 'polished';
 import { css } from '@emotion/css';
 import { Palette } from '../../design/color';
 import { GameState, isSpyMaster, Team } from '../../model';
 import { beginAt } from '../../design/responsive';
-
-const StyledContainer = styled.div<{ color: Team }>`
-  display: flex;
-  border-radius: 0.5rem;
-  padding: 0.5rem;
-  margin: 1rem;
-  font-size: 0.5rem;
-  flex-direction: column;
-  color: ${Palette.contrast};
-  background-color: ${({ color }) =>
-    color === 'Blue' ? Palette.blue : Palette.red};
-
-  h2 {
-    margin: 0;
-  }
-
-  ul {
-    display: flex;
-    flex-wrap: wrap;
-    padding-left: 0.75rem;
-  }
-
-  li {
-    margin: 0.25rem 0;
-    padding-right: 0.5rem;
-  }
-
-  button {
-    padding: 0.5rem;
-    border-radius: 0.5rem;
-    font-size: 0.5rem;
-    background-color: ${Palette.neutral};
-    cursor: pointer;
-    :hover {
-      background-color: ${lighten(0.1, Palette.neutral)};
-    }
-    ${beginAt(375)} {
-      font-size: .75rem;
-    }
-
-    ${beginAt(768)} {
-      font-size: 1rem;
-    }
-  }
-  
-  ${beginAt(375)} {
-    font-size: .75rem;
-  }
-
-  ${beginAt(768)} {
-    font-size: 1rem;
-  }
-}
-`;
 
 export interface PlayerListProps {
   players: GameState['players'];
@@ -83,7 +28,7 @@ export const PlayerList: FC<PlayerListProps> = ({
     onJoin(name, team, spyMaster ? '' : null); // TODO: figure out how to pass secret
   }, [team, spyMaster, onJoin]);
   return (
-    <StyledContainer color={team}>
+    <div className={styleContainer(team)}>
       <div
         className={css`
           align-self: flex-start;
@@ -111,9 +56,63 @@ export const PlayerList: FC<PlayerListProps> = ({
           {spyMaster ? 'Join as Spymaster' : 'Join as Operative'}
         </button>
       )}
-    </StyledContainer>
+    </div>
   );
 };
+
+function styleContainer(color: Team) {
+  return css`
+    display: flex;
+    border-radius: 0.5rem;
+    padding: 0.5rem;
+    margin: 1rem;
+    font-size: 0.5rem;
+    flex-direction: column;
+    color: ${Palette.contrast};
+    background-color: ${color === 'Blue' ? Palette.blue : Palette.red};
+
+    h2 {
+      margin: 0;
+    }
+
+    ul {
+      display: flex;
+      flex-wrap: wrap;
+      padding-left: 0.75rem;
+    }
+
+    li {
+      margin: 0.25rem 0;
+      padding-right: 0.5rem;
+    }
+
+    button {
+      padding: 0.5rem;
+      border-radius: 0.5rem;
+      font-size: 0.5rem;
+      background-color: ${Palette.neutral};
+      cursor: pointer;
+      :hover {
+        background-color: ${lighten(0.1, Palette.neutral)};
+      }
+      ${beginAt(375)} {
+        font-size: 0.75rem;
+      }
+
+      ${beginAt(768)} {
+        font-size: 1rem;
+      }
+    }
+
+    ${beginAt(375)} {
+      font-size: 0.75rem;
+    }
+
+    ${beginAt(768)} {
+      font-size: 1rem;
+    }
+  `;
+}
 
 function getPlayerNames(
   players: GameState['players'],
