+++ app/pages/_app.tsx
@@ -1,7 +1,8 @@
 import { AppProps } from 'next/app';
 import { FC } from 'react';
-import { Global } from '@emotion/react';
-import { GlobalStyle, Layout } from '../design/layout';
+import { css, Global } from '@emotion/react';
+import reset from 'emotion-reset';
+import { Layout } from '../design/layout';
 
 const App: FC<AppProps> = ({ Component, pageProps }) => (
   <>
@@ -11,5 +12,27 @@ const App: FC<AppProps> = ({ Component, pageProps }) => (
     </Layout>
   </>
 );
+
+export const GlobalStyle = css`
+  ${reset};
+  body {
+    font-family: Montserrat, 'Arial CE', Arial, sans-serif;
+    & * {
+      box-sizing: border-box;
+    }
+  }
+  button,
+  input {
+    border-style: solid;
+  }
+  h1 {
+    font-size: 2rem;
+    margin: 2rem 0;
+  }
+  h2 {
+    font-size: 1.5rem;
+    margin: 1.5rem;
+  }
+`;
 
 export default App;
