+++ app/pages/index.tsx
@@ -1,8 +1,10 @@
 import { FC } from 'react';
 import { GetServerSideProps } from 'next';
 import Link from 'next/link';
+import { css } from '@emotion/css';
 import { NewGame } from '../components/NewGame';
-import { ContentContainer } from '../components/ContentContainer';
+import { Palette } from '../design/color';
+import { Breakpoints } from '../design/responsive';
 
 interface HomeProps {
   game_name: string;
@@ -10,7 +12,7 @@ interface HomeProps {
 }
 
 const Home: FC<HomeProps> = ({ API_URL, game_name }) => (
-  <ContentContainer>
+  <div className={styleContent}>
     <div>
       <h2>create a new game</h2>
       <NewGame API_URL={API_URL} initialName={game_name} />
@@ -18,8 +20,26 @@ const Home: FC<HomeProps> = ({ API_URL, game_name }) => (
     <h2>
       or <Link href="/game">join an existing one</Link>
     </h2>
-  </ContentContainer>
+  </div>
 );
+
+const { neutral, blue, contrast } = Palette;
+
+export const styleContent = css`
+  background-color: ${neutral};
+  display: flex;
+  flex-direction: column;
+  margin: auto;
+  padding: 1rem;
+  border-radius: 1rem;
+  box-shadow: 0 0 2px 1px ${blue};
+  color: ${contrast};
+  max-width: ${Breakpoints.tabletPortrait}px;
+  text-align: center;
+  * {
+    color: ${contrast};
+  }
+`;
 
 export const getServerSideProps: GetServerSideProps<HomeProps> = async () => {
   const API_URL = process.env.API_URL as string;
